+++ src/main/java/com/example/FastLane/Academy/controller/EnrollmentController.java
@@ -3,8 +3,10 @@ package com.example.FastLane.Academy.controller;
 import com.example.FastLane.Academy.dto.EnrollmentDTO;
 import com.example.FastLane.Academy.dto.ResponseDTO;
 import com.example.FastLane.Academy.service.EnrollmentService;
+import com.example.FastLane.Academy.util.SessionUtil;
 import com.example.FastLane.Academy.util.VarList;
 import com.example.FastLane.Academy.service.AdminApprovalService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -23,7 +25,9 @@ public class EnrollmentController {
 
     // Enroll Student
     @PostMapping("/enroll")
-    public ResponseEntity<ResponseDTO> enrollStudent(@RequestBody EnrollmentDTO enrollmentDTO) {
+    public ResponseEntity<ResponseDTO> enrollStudent(@RequestBody EnrollmentDTO enrollmentDTO, HttpSession session)
+    {
+
 
         ResponseDTO response =
                 enrollmentService.enrollStudent(enrollmentDTO);
@@ -39,7 +43,7 @@ public class EnrollmentController {
     // View Courses By Student
     @GetMapping("/student/{studentId}")
     public ResponseEntity<ResponseDTO> getCoursesByStudent(
-            @PathVariable String studentId) {
+            @PathVariable String studentId,HttpSession session) {
 
         ResponseDTO response =
                 enrollmentService.getCoursesByStudent(studentId);
@@ -51,8 +55,11 @@ public class EnrollmentController {
     // View Students By Course
     @GetMapping("/course/{courseId}")
     public ResponseEntity<ResponseDTO> getStudentsByCourse(
-            @PathVariable String courseId) {
-
+            @PathVariable String courseId, HttpSession session) {
+        if (!SessionUtil.isRole(session, "ADMIN")) {
+            return ResponseEntity.status(403)
+                    .body(new ResponseDTO(VarList.UNAUTHORIZED, "Admin access only", null));
+        }
         ResponseDTO response =
                 enrollmentService.getStudentsByCourse(courseId);
 
@@ -63,7 +70,12 @@ public class EnrollmentController {
     // Remove Enrollment
     @PutMapping("/remove/{enrollmentId}")
     public ResponseEntity<ResponseDTO> removeEnrollment(
-            @PathVariable String enrollmentId) {
+            @PathVariable String enrollmentId,HttpSession session)
+    {
+        if (!SessionUtil.isRole(session, "ADMIN")) {
+            return ResponseEntity.status(403)
+                    .body(new ResponseDTO(VarList.UNAUTHORIZED, "Admin access only", null));
+        }
 
         ResponseDTO response =
                 enrollmentService.removeEnrollment(enrollmentId);
@@ -80,7 +92,12 @@ public class EnrollmentController {
 
     // Admin: approve enrollment access manually (edge case) direct access
     @PutMapping("/approve/{enrollmentId}")
-    public ResponseEntity<ResponseDTO> approveEnrollment(@PathVariable String enrollmentId) {
+    public ResponseEntity<ResponseDTO> approveEnrollment(@PathVariable String enrollmentId,HttpSession session)
+    {
+        if (!SessionUtil.isRole(session, "ADMIN")) {
+            return ResponseEntity.status(403)
+                    .body(new ResponseDTO(VarList.UNAUTHORIZED, "Admin access only", null));
+        }
         ResponseDTO response = adminApprovalService.approveEnrollmentById(enrollmentId);
         return ResponseEntity.status(HttpStatus.ACCEPTED).body(response);
     }
@@ -89,7 +106,12 @@ public class EnrollmentController {
     @PutMapping("/reject/{enrollmentId}")
     public ResponseEntity<ResponseDTO> rejectEnrollment(
             @PathVariable String enrollmentId,
-            @RequestParam String reason) {
+            @RequestParam String reason, HttpSession session)
+    {
+        if (!SessionUtil.isRole(session, "ADMIN")) {
+            return ResponseEntity.status(403)
+                    .body(new ResponseDTO(VarList.UNAUTHORIZED, "Admin access only", null));
+        }
         ResponseDTO response = adminApprovalService.rejectEnrollmentById(enrollmentId, reason);
         return ResponseEntity.status(HttpStatus.ACCEPTED).body(response);
     }
